Extract notices
